% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_patient.R
\name{convert_patient}
\alias{convert_patient}
\title{Filter and summarize IHC cell level data into sample level}
\usage{
convert_patient(data, patient, core = "A")
}
\arguments{
\item{data}{a dataframe which contains the spatial score, which is returned by the spatial_score function}

\item{patient}{The xlsx file with path should contain the patient information, related slide ID and low quality cores.}

\item{core}{The core used to calculate patient label data}
}
\value{
a dataframe which contains the marker signal for CD30, CD4, CD68, CD20 and PD1
}
\description{
The column names of a given data are hard coded, further changes are needed if data format is changed
}
\author{
Yifan Yin, Aixiang Jiang
}
