% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_markerFolders.R
\name{merge_markerFolders}
\alias{merge_markerFolders}
\title{Merge IHC data into a single data frame}
\usage{
merge_markerFolders(
  path,
  markerFolders = c("CD30", "CD4", "CD68", "CD20", "PD1", "CXCR5")
)
}
\arguments{
\item{path}{the location where the IHC data is stored. Under the path, there should be 6 separate folders which contain the IHC output
for each markers. The name for each subfolder is the marker name e.g. CD30, CD20 and etc.}

\item{markerFolders}{the IHC raw data folder needed for RHL4S, you could have other subfolders, but at least these subfolders should be included}
}
\value{
a dataframe which contains the marker signal for CD30, CD4, CD68, CD20 and PD1
}
\description{
Please notice that the subfolders' names and content are hard coded, further changes are needed if data names and formats are changed
}
\author{
Yifan Yin, Aixiang Jiang
}
