% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_RHL4S.R
\name{predict_RHL4S}
\alias{predict_RHL4S}
\title{A function to get RHL4S calls}
\usage{
predict_RHL4S(
  newdat,
  varsIn = c("CXCR5_HRS_spatial_score", "PD1_CD4_spatial_score", "Mac_spatial_score",
    "CXCR5_B_spatial_score")
)
}
\arguments{
\item{newdat}{A new IHC data frame, samples are in columns, and spatial scores are in rows.
Notice that data are already pre-processed}

\item{varsIn}{The model variable list}
}
\value{
A data frame with LPS score, Empirical Bayesian probabilities for two groups and classification
}
\description{
This is the function to calculate RHL4S LPS (Linear Predictor Score) scores and empiracal Bayes' probs,
and get RHL4S calls for a new IHC data set. The required prior information is:
4 selected RHL4S spatial scores and their weights, LPS score means and sds for RHL4S POS and NEG groups
}
\details{
Make sure the new IHC data is comparable to the BCCA IHC data, if not, calibration is required before calling this function
}
\author{
Aixiang Jiang
}
\keyword{IHC}
\keyword{LPS,}
\keyword{RHL4S,}
