% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHL4Scalls.R
\name{RHL4Scalls}
\alias{RHL4Scalls}
\title{A wrap-up function to get RHL4S calls from IHC raw data}
\usage{
RHL4Scalls(datapath = NULL, patientFile = NULL, patientSheet = "A")
}
\arguments{
\item{datapath}{This data folder should contain 6 subfolder for each marker separately: CD30, CD4, CD68, CD20,PD1,CXCR5.}

\item{patientFile}{The xlsx or xls file with path should contain the patient information, related slide ID and low quality cores.}

\item{patientSheet}{The sheet name in patientFile to indicate the core used to calculate patient label data}
}
\value{
A data frame with patient level spatial score, LPS score, Empirical Bayesian probabilities for two groups and classification
}
\description{
This is the function to integrate IHC data for the given markers, calculate spatial score, filer cells, summarize to patient level data,
  and calculate RHL4S LPS (Linear Predictor Score) scores and empiracal Bayes' probabilities, and classification based on probability cutoff
}
\details{
Make sure the new IHC data is comparable to the BCCA IHC data, if not, calibration is required
}
\author{
Aixiang Jiang
}
\keyword{IHC}
\keyword{LPS,}
\keyword{RHL4S,}
