% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_score.R
\name{spatial_score}
\alias{spatial_score}
\title{Calculate cell level spatial scores}
\usage{
spatial_score(data)
}
\arguments{
\item{data}{a dataframe which includes the merged results for CD30, CD4, CD68, CD20, CXCR5 and PD1}
}
\value{
a dataframe which contains the spatial score at cell level
}
\description{
The column names of a given data are hard coded, further changes are needed if data format is changed
}
\author{
Alex Xu, Yifan Yin, Aixiang Jiang
}
